+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- User Roles interface
+
 ## [2.13.0] - 2022-03-04
 
 ### Added